requirements.txt
Django==2.2.4
psycopg2-binary==2.8.3
djangorestframework==3.9.0
Markdown==3.1.1
django-filter==2.2.0
django-debug-toolbar==2.0
social-auth-app-django==3.1.0
django-simple-history==2.7.3
boto3==1.9.214
django-extensions==2.2.1
tqdm==4.34.0
flake8==3.7.8
drf-yasg==1.15.1
## The following requirements were added by pip freeze:
botocore==1.12.214
certifi
chardet==3.0.4
coreapi==2.3.3
coreschema==0.0.4
defusedxml==0.6.0
docutils==0.15.2
entrypoints==0.3
idna==2.8
inflection==0.3.1
itypes==1.1.0
Jinja2==2.10.1
jmespath==0.9.4
MarkupSafe==1.1.1
mccabe==0.6.1
oauthlib==3.1.0
pycodestyle==2.5.0
pyflakes==2.1.1
PyJWT==1.7.1
python-dateutil==2.8.0
python3-openid==3.1.0
pytz==2019.2
requests==2.22.0
requests-oauthlib==1.2.0
ruamel.yaml==0.16.5
ruamel.yaml.clib==0.1.2
s3transfer==0.2.1
six==1.12.0
social-auth-core==3.2.0
sqlparse==0.3.0
uritemplate==3.0.0
urllib3==1.25.3
watchtower==0.6.0
zappa
zappa_django_utils
django-s3-storage
psycopg2
